Module Tezos_rpc_http_server

module RPC_logging : sig ... end

This module is internal to the RPC layer. It provides logging interface required by resto HTTP library and translates its log messages to events in Tezos' internal logging system. It is not intended to be used for any other purpose. For your own logging use Internal_event.Simple instead.

module RPC_server : sig ... end

Typed RPC services: server implementation.