Module Tezos_rpc_http_server.RPC_logging

This module is internal to the RPC layer. It provides logging interface required by resto HTTP library and translates its log messages to events in Tezos' internal logging system. It is not intended to be used for any other purpose. For your own logging use Internal_event.Simple instead.

val debug : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val log_info : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val log_notice : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val warn : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val log_error : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val lwt_debug : ('a, Stdlib.Format.formatter, unit, unit Lwt.t) Stdlib.format4 -> 'a

val lwt_log_info : ('a, Stdlib.Format.formatter, unit, unit Lwt.t) Stdlib.format4 -> 'a

val lwt_log_notice : ('a, Stdlib.Format.formatter, unit, unit Lwt.t) Stdlib.format4 -> 'a

val lwt_warn : ('a, Stdlib.Format.formatter, unit, unit Lwt.t) Stdlib.format4 -> 'a

val lwt_log_error : ('a, Stdlib.Format.formatter, unit, unit Lwt.t) Stdlib.format4 -> 'a