Module Tezos_rpc_http_server__RPC_server.Acl

include module type of Resto_acl.Acl

type chunk_matcher = Resto_acl__Acl.chunk_matcher =

| Literal of string

| Wildcard

type path_matcher = Resto_acl__Acl.path_matcher =

| Exact of chunk_matcher list

| FollowedByAnySuffix of chunk_matcher list

type meth_matcher = Resto_acl__Acl.meth_matcher =

| Exact of Resto.meth

| Any

type matcher = Resto_acl__Acl.matcher = {

meth : meth_matcher;

path : path_matcher;

}

val parse : string -> matcher

type t = Resto_acl__Acl.t =

| Allow_all of {

except : matcher list;

}

| Deny_all of {

except : matcher list;

}

val allowed : t -> meth:Resto.meth -> path:string list -> bool

type policy

A policy for the whole RPC server is a set of access control lists for the different addresses that the RPC server listens to. It is represented as an association list mapping listening addresses (in string literal forms) to deny/access access policies.

val default : t

Default ACL policy in case none is defined in configuration. For the sake of backwards compatibility it allows access to all resources. This will probably change in the future so that access to sensitive resources is restricted.

val put_policy : (Tezos_base__TzPervasives.P2p_point.Id.addr_port_id * t) -> policy -> policy

Add an ACL for given address into the policy. Overrides previously existing policy for that address if any.

val empty_policy : policy

Empty ACL policy allows access to all endpoints. Currently it's the same as default below, but that will likely change in the future, therefore it's better to use default rather than this value. It's mainly intended for testing.

val default_policy : policy

This is the default policy. Currently equivalent to empty above, but that will likely change at some point in the future.

val policy_encoding : policy Tezos_base__TzPervasives.Data_encoding.t

val policy_to_string : policy -> string

Returns the JSON representation of the policy.

val find_policy : policy -> string -> t option

find_policy policy address looks for the address within the policy and returns corresponding access control list.

An ACL is considered matching if its corresponding host-name part matches the host-name part of the address and either:

its corresponding port also matches address's port OR

its corresponding address does not mention any port at all.

The first ACL whose corresponding address matches these criteria is returned.

val matcher_to_string : matcher -> string

Returns string representation of a given matcher. Useful for testing.